feat: add likes and follows table with models update

+++ database/migrations/2021_11_10_133853_create_posts_table.php
@@ -19,7 +19,7 @@ class CreatePostsTable extends Migration
             $table->string('img_url');
             $table->string('min_img_url');
             $table->string('tags')->nullable();
-            $table->integer('author_id');
+            $table->bigInteger('author_id');
             $table->timestamps();
         });
     }
